+++ libs/three.extra.js
@@ -2525,6 +2525,7 @@ THREE.OrbitControls = function ( object, domElement ) {
 		switch ( event.pointerType ) {
 
 			case 'mouse':
+			case 'pen':
 				onMouseDown( event );
 				break;
 
@@ -2541,6 +2542,7 @@ THREE.OrbitControls = function ( object, domElement ) {
 		switch ( event.pointerType ) {
 
 			case 'mouse':
+			case 'pen':
 				onMouseMove( event );
 				break;
 
@@ -2557,6 +2559,7 @@ THREE.OrbitControls = function ( object, domElement ) {
 		switch ( event.pointerType ) {
 
 			case 'mouse':
+			case 'pen':
 				onMouseUp( event );
 				break;
 
@@ -3094,7 +3097,6 @@ THREE.TransformControls = function ( camera, domElement ) {
 
 	{
 
-		domElement.addEventListener( "touchstart", onTouchStart, false );
 		domElement.addEventListener( "pointerdown", onPointerDown, false );
 		domElement.addEventListener( "pointermove", onPointerHover, false );
 		scope.domElement.ownerDocument.addEventListener( "pointerup", onPointerUp, false );
@@ -3103,7 +3105,6 @@ THREE.TransformControls = function ( camera, domElement ) {
 
 	this.dispose = function () {
 
-		domElement.removeEventListener( "touchstart", onTouchStart );
 		domElement.removeEventListener( "pointerdown", onPointerDown );
 		domElement.removeEventListener( "pointermove", onPointerHover );
 		scope.domElement.ownerDocument.removeEventListener( "pointermove", onPointerMove );
@@ -3587,6 +3588,7 @@ THREE.TransformControls = function ( camera, domElement ) {
 		switch ( event.pointerType ) {
 
 			case 'mouse':
+			case 'pen':
 				scope.pointerHover( getPointer( event ) );
 				break;
 
@@ -3598,6 +3600,7 @@ THREE.TransformControls = function ( camera, domElement ) {
 
 		if ( ! scope.enabled ) return;
 
+		scope.domElement.style.touchAction = 'none'; // disable touch scroll
 		scope.domElement.ownerDocument.addEventListener( "pointermove", onPointerMove, false );
 
 		scope.pointerHover( getPointer( event ) );
@@ -3617,17 +3620,10 @@ THREE.TransformControls = function ( camera, domElement ) {
 
 		if ( ! scope.enabled ) return;
 
+		scope.domElement.style.touchAction = '';
 		scope.domElement.ownerDocument.removeEventListener( "pointermove", onPointerMove, false );
 
 		scope.pointerUp( getPointer( event ) );
-
-	}
-
-	function onTouchStart( event ) {
-
-		if ( scope.enabled === false ) return;
-
-		event.preventDefault(); // prevent scrolling
 
 	}