+++ changes.md
@@ -26,6 +26,7 @@
 16. Update jquery to 3.5.1, openui5 to 1.82.2
 17. Add support of log2 scale for axes drawing, can be extended for arbitrary log base
 18. Allow to move axis title to opposite position
+19. Fix zooming in color palette
 
 
 ## Changes in 5.9.0